MNT-20206 Improper Resource Shutdown or Release CWE ID 404
The application fails to release (or incorrectly releases) a system resource before it is made available for re-use. This condition often occurs with resources such as database connections or file handles. Most unreleased resource issues result in general software reliability problems, but if an attacker can intentionally trigger a resource leak, it may be possible to launch a denial of service attack by depleting the resource pool.

- use try-with-resources pattern to ensure bufferedReader is closed